Adjustments following Mike's comments
