+++ src/tests/Microsoft.PowerFx.Connectors.Tests/OpenApiParserTests.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Net.Http;
 using System.Threading;
@@ -41,48 +39,7 @@ namespace Microsoft.PowerFx.Connectors.Tests
             Assert.Equal("Conversations (CLU) (2022-05-01)", function.Summary);
             Assert.Equal("/apim/cognitiveservicestextanalytics/{connectionId}/language/:analyze-conversations", function.OperationPath);
             Assert.Equal(HttpMethod.Post, function.HttpMethod);
-        }
-
-        [Fact(Skip = "Need files from AAPT-connector and PowerPlatformConnectors projects")]
-        public void TestAllConnectors()
-        {
-            int i = 0;
-            int j = 0;
-
-            // Store results for analysis
-            using StreamWriter writer = new StreamWriter(@"c:\temp\out.txt", append: false);
-
-            foreach (string swaggerFile in Directory.EnumerateFiles(@"C:\Data\AAPT-connectors\src", "apidefinition*swagger*json", new EnumerationOptions() { RecurseSubdirectories = true })
-                                    .Union(Directory.EnumerateFiles(@"C:\Data\PowerPlatformConnectors", "apidefinition*swagger*json", new EnumerationOptions() { RecurseSubdirectories = true })))
-            {
-                i++;
-                try
-                {
-                    OpenApiDocument doc = Helpers.ReadSwagger(swaggerFile);
-
-                    // Check we can get the functions
-                    IEnumerable<ConnectorFunction> functions = OpenApiParser.GetFunctions(doc);
-
-                    var config = new PowerFxConfig();
-                    using var client = new PowerPlatformConnectorClient("firstrelease-001.azure-apim.net", "839eace6-59ab-4243-97ec-a5b8fcc104e4", "72c42ee1b3c7403c8e73aa9c02a7fbcc", () => "Some JWT token")
-                    {
-                        SessionId = "ce55fe97-6e74-4f56-b8cf-529e275b253f"
-                    };
-
-                    // Check we can add the service (more comprehensive test)
-                    config.AddService("Connector", doc, client);
-
-                    writer.WriteLine($"{swaggerFile}: OK - functions: {functions.Count()}");
-                }
-                catch (Exception ex)
-                {
-                    writer.WriteLine($"{swaggerFile}: Exception {ex.GetType().Name} - {ex.Message}");
-                    j++;
-                }
-            }
-
-            writer.WriteLine($"Total: {i} - Exceptions: {j}");
-        }
+        }       
 
         [Fact]
         public void ACSL_Load()
